implemented level clear conditions and methods

+++ test/src/org/academiadecodigo/apiores/test/Test.java
@@ -17,6 +17,12 @@ public class Test {
 
     private boolean dead = false;
 
+    // trying to implement level cleared conditions
+
+    private boolean cleared = false;
+    private Rectangle levelObjective;
+
+
 
     public Test() {
 
@@ -24,17 +30,31 @@ public class Test {
         Shape rec = new Rectangle(10, 10, width , height);
         canvas.show(rec);
         rectangle = new Rectangle(480, 450, 50, 50);
+
+        // when rectangle and levelObjective share the same position the level clears
+
+        levelObjective = new Rectangle(0, 600, 50, 50);
     }
 
     public void start() throws InterruptedException {
 
         while(true) {
+
+            if(cleared){
+                break;
+            }
+
             rectangle.delete();
+            levelObjective.delete();
             rectangle = new Rectangle(480, 450, 50, 50);
+            levelObjective = new Rectangle(600, 10, 50, 50);
 
             KeyListener keyboard = new KeyListener(rectangle, 20);
             rectangle.setColor(Color.BLUE);
             rectangle.fill();
+
+            levelObjective.setColor(Color.ORANGE);
+            levelObjective.fill();
 
 
             createCars(firstTrack, 15, 200, 140);
@@ -47,28 +67,35 @@ public class Test {
             borderLeft.setColor(Color.BLUE);
             borderLeft.fill();
             borderRight.fill();
-            while (!dead) {
+            while (!dead && !cleared) {
                 for (MoveCars car : firstTrack) {
                     car.moveCarLeft();
                     checkDead(car);
+                    checkCleared();
 
                 }
                 for (MoveCars car : secondTrack) {
                     car.moveCarLeft();
                     checkDead(car);
+                    checkCleared();
 
                 }
                 for (MoveCars car : thirdTrack) {
                     car.moveCarLeft();
                     checkDead(car);
+                    checkCleared();
                 }
                 Thread.sleep(75);
 
 
             }
+
+            if(dead){
             rectangle.setColor(Color.RED);
+            }
 
             while (dead) {
+
                 System.out.println();
             }
 
@@ -115,6 +142,15 @@ public class Test {
         }
     }
 
+    public void checkCleared(){
+        if(rectangle.getX() == levelObjective.getX() && rectangle.getY() == levelObjective.getY()){
+
+            cleared = true;
+        }
+
+    }
+
+
     public boolean isDead() {
         return dead;
     }
@@ -126,6 +162,8 @@ public class Test {
         deleteCars(firstTrack);
         deleteCars(secondTrack);
         deleteCars(thirdTrack);
+        levelObjective.delete();
+
         dead=false;
 
     }
